:truck: Rename DormantGitHubUser to DormantGitHubUserService
We've renamed the 'DormantGitHubUser' module to 'DormantGitHubUserService' to better align with the standard nomenclature in the existing codebase. This change has also been reflected in the import statements of its dependent scripts. We've also added a new test file for the renamed service.
